removed a geometryerror, added 'find' to util.py
